+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Changed
+- Suppressed automatic camera zoom/follow when changing the origin or destination in the mission planning window.
 
 ## [1.1.0] - 2026-05-16
 ### Changed